% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{return_to_cumret}
\alias{return_to_cumret}
\title{calculate continius cummulated returns from returns}
\usage{
return_to_cumret(data_xts)
}
\description{
calculate continius cummulated returns from returns
}
